requirements.txt
crawl4ai>=0.6.2
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pytest>=8.0.0
httpx>=0.27.0 